Refactor VLA loader to use PyTorch dataloader for improved code readability and performance
